server/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.6.3
requests==2.31.0
flask-cors==4.0.0
psycopg2==2.9.9
gunicorn==21.2.0
flask_caching==2.3.0
redis==5.0.4
